FIRST COMMIT
created engine, gameobject, and scene metaclasses to be used for all 
objects. Also added a testGame folder which will be used for testing all 
features, and can be duplicated to make a new game. To do now will be 
assigning game objects positions and sprites then displaying the with 
pygame, after that allow for scene switching and make sure that does not 
cause issues.